[Chore] Restructure Project Folder
