+++ .sequelizerc
@@ -3,6 +3,6 @@ require('@babel/register')
 module.exports = {
   "config": path.resolve('./src/database', 'config.js'),
   "models-path": path.resolve('./src/models'),
-  "seeders-path": path.resolve('./src/seeders'),
-  "migrations-path": path.resolve('./src/migrations')
+  "seeders-path": path.resolve('./src/database/seeders'),
+  "migrations-path": path.resolve('./src/database/migrations')
 };
