--- src/modules/auth/routes.js
@@ -1,7 +1,0 @@
-import express from 'express';
-import auth from './auth.routes';
-
-const router = express.Router();
-router.use('/auth', auth);
-
-export default router;
