+++ src/routes/index.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import router from './api';
+
+const apiRouter = express.Router();
+
+apiRouter.use('/api', router);
+
+export default apiRouter;
